# frontend/app/swagger/api-schema.yaml
openapi: 3.0.3
info:
  title: Recipe Cookbook API
  version: 1.0.0
  description: A recipe cookbook API with detailed cooking instructions and ingredient measurements
paths:
  /:
    get:
      operationId: home
      description: Home page displaying all recipes with 90s styling
      tags:
        - web
      responses:
        "200":
          description: HTML page with recipe list
          content:
            text/html:
              schema:
                type: string

  /recipes/{id}/:
    get:
      operationId: recipe_detail
      description: Recipe detail page with full cooking instructions (WARNING - Vulnerable to SQL injection for educational purposes)
      tags:
        - web
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: Recipe ID
          required: true
      responses:
        "200":
          description: HTML page with recipe details
          content:
            text/html:
              schema:
                type: string

  /api/user/create/:
    post:
      operationId: user_create_create
      description: Create a new user in the system.
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: ""

  /api/user/me/:
    get:
      operationId: user_me_retrieve
      description: Manage the authenticated user.
      tags:
        - user
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: ""
    put:
      operationId: user_me_update
      description: Manage the authenticated user.
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: ""
    patch:
      operationId: user_me_partial_update
      description: Manage the authenticated user.
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedUserRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: ""

  /api/user/token/:
    post:
      operationId: user_token_create
      description: Create a new auth token for user.
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthTokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
          description: ""

  /api/recipe/recipes/:
    get:
      operationId: recipe_recipes_list
      description: View for manage recipe APIs.
      parameters:
        - in: query
          name: ingredients
          schema:
            type: string
          description: Comma separated list of ingredient IDs to filter
        - in: query
          name: tags
          schema:
            type: string
          description: Comma separated list of tag IDs to filter
      tags:
        - recipe
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
          description: ""
    post:
      operationId: recipe_recipes_create
      description: View for manage recipe APIs.
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeDetailRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
          description: ""

  /api/recipe/recipes/{id}/:
    get:
      operationId: recipe_recipes_retrieve
      description: View for manage recipe APIs.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this recipe.
          required: true
      tags:
        - recipe
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
          description: ""
    put:
      operationId: recipe_recipes_update
      description: View for manage recipe APIs.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this recipe.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeDetailRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
          description: ""
    patch:
      operationId: recipe_recipes_partial_update
      description: View for manage recipe APIs.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this recipe.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedRecipeDetailRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
          description: ""
    delete:
      operationId: recipe_recipes_destroy
      description: View for manage recipe APIs.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this recipe.
          required: true
      tags:
        - recipe
      responses:
        "204":
          description: No response body

  /api/recipe/recipes/{id}/upload-image/:
    post:
      operationId: recipe_recipes_upload_image_create
      description: Upload an image to recipe.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this recipe.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeImageRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeImage"
          description: ""

  /api/recipe/ingredients/:
    get:
      operationId: recipe_ingredients_list
      description: Manage ingredients in the database.
      parameters:
        - in: query
          name: assigned_only
          schema:
            type: integer
            enum:
              - 0
              - 1
          description: Filter by items assigned to recipes.
      tags:
        - recipe
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
          description: ""

  /api/recipe/ingredients/{id}/:
    put:
      operationId: recipe_ingredients_update
      description: Manage ingredients in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this ingredient.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
          description: ""
    patch:
      operationId: recipe_ingredients_partial_update
      description: Manage ingredients in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this ingredient.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedIngredientRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
          description: ""
    delete:
      operationId: recipe_ingredients_destroy
      description: Manage ingredients in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this ingredient.
          required: true
      tags:
        - recipe
      responses:
        "204":
          description: No response body

  /api/recipe/tags/:
    get:
      operationId: recipe_tags_list
      description: Manage tags in the database.
      parameters:
        - in: query
          name: assigned_only
          schema:
            type: integer
            enum:
              - 0
              - 1
          description: Filter by items assigned to recipes.
      tags:
        - recipe
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
          description: ""

  /api/recipe/tags/{id}/:
    put:
      operationId: recipe_tags_update
      description: Manage tags in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this tag.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
          description: ""
    patch:
      operationId: recipe_tags_partial_update
      description: Manage tags in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this tag.
          required: true
      tags:
        - recipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchedTagRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
          description: ""
    delete:
      operationId: recipe_tags_destroy
      description: Manage tags in the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          description: A unique integer value identifying this tag.
          required: true
      tags:
        - recipe
      responses:
        "204":
          description: No response body

components:
  schemas:
    AuthToken:
      type: object
      description: Serializer for the user auth token.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthTokenRequest:
      type: object
      description: Serializer for the user auth token.
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - email
        - password

    Ingredient:
      type: object
      description: Serializer for ingredients.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        amount:
          type: string
          description: Amount of ingredient (e.g., "400", "2")
        unit:
          type: string
          description: Unit of measurement (e.g., "g", "tbsp", "pieces")
      required:
        - id
        - name

    IngredientRequest:
      type: object
      description: Serializer for ingredients.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        amount:
          type: string
        unit:
          type: string
      required:
        - name

    PatchedIngredientRequest:
      type: object
      description: Serializer for ingredients.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        amount:
          type: string
        unit:
          type: string

    PatchedRecipeDetailRequest:
      type: object
      description: Serializer for recipe detail view.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        time_minutes:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        link:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRequest"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientRequest"
        description:
          type: string
          description: Step-by-step cooking instructions

    PatchedTagRequest:
      type: object
      description: Serializer for tags.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255

    PatchedUserRequest:
      type: object
      description: Serializer for the user object.
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        password:
          type: string
          writeOnly: true
          minLength: 5
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 255

    Recipe:
      type: object
      description: Serializer for recipes.
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        time_minutes:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        link:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
      required:
        - id
        - price
        - time_minutes
        - title

    RecipeDetail:
      type: object
      description: Serializer for recipe detail view with step-by-step instructions.
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
        time_minutes:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        link:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        description:
          type: string
          description: Step-by-step cooking instructions for the recipe
      required:
        - id
        - price
        - time_minutes
        - title

    RecipeDetailRequest:
      type: object
      description: Serializer for recipe detail view.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        time_minutes:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        price:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
        link:
          type: string
          maxLength: 255
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRequest"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/IngredientRequest"
        description:
          type: string
          description: Step-by-step cooking instructions
      required:
        - price
        - time_minutes
        - title

    RecipeImage:
      type: object
      description: Serializer for uploading images to recipes.
      properties:
        id:
          type: integer
          readOnly: true
        image:
          type: string
          format: uri
          nullable: true
      required:
        - id
        - image

    RecipeImageRequest:
      type: object
      description: Serializer for uploading images to recipes.
      properties:
        image:
          type: string
          format: binary
          nullable: true
      required:
        - image

    Tag:
      type: object
      description: Serializer for tags.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
      required:
        - id
        - name

    TagRequest:
      type: object
      description: Serializer for tags.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
      required:
        - name

    User:
      type: object
      description: Serializer for the user object.
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        name:
          type: string
          maxLength: 255
      required:
        - email
        - name

    UserRequest:
      type: object
      description: Serializer for the user object.
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        password:
          type: string
          writeOnly: true
          minLength: 5
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 255
      required:
        - email
        - name
        - password